encoding.rb fix for Ruby 1.8.7 array discrepancy
There is a fundamental difference between Ruby 1.8.7's Array() and [] wrappers.  The first, Array(), if given a string with a newline character (\n) will split the string at the newline; whereas the latter version (i.e. []) will give the expected result of keeping the string in tact as a single entity.

This fix is necessary since a Quoted-Printable RFC-1342 encoded-word can have newline characters within the encoded text and when it is split apart it is not seen as an RFC-1342 encoded-word.